Remove unnecessary shell-echo disable
This was recommended by Copilot, "in case" a shell was running in a
debug mode ... but then a subsequent Copilot run complained about it.
After researching some more, I've come to the conclusion that is is the
caller's responsibility to either pass a secret, e.g. `secret.FOO`,
for which GitHub has automatic masking enforced should it be echoed, or,
for the caller to prefix any non-secret sensitive value used as an input
like, `::add-mask::VALUE`.

Our expected usage for GHCR will be `secrets.GITHUB_TOKEN` so we are
covered there.

Signed-off-by: Eric Searcy <eric@linuxfoundation.org>