Always clobber chart version with passed value
Currently, this shared action requires a chart version, which the
callers all currently extract FROM Chart.yaml, and then this passed
version is used in the shared action for file names. (Which is an odd
choice, since THIS action could have just extracted it, but oh well.)

Anyhow, by always doing an inline edit of Chart.yaml to set/replace the
Chart's `version` attribute with the passed version, we will also now
allow the callers to choose a different mechanism (namely, the release
tag), to set the chart version.

Signed-off-by: Eric Searcy <eric@linuxfoundation.org>